+++ sdk/src/main/java/co/omise/android/models/SourceType.kt
@@ -24,6 +24,15 @@ sealed class SourceType(
     object PointsCiti : SourceType("points_citi")
     object PayNow : SourceType("paynow")
     object PromptPay : SourceType("promptpay")
+    object AlipayCn : SourceType("alipay_cn")
+    object AlipayHk : SourceType("alipay_hk")
+    object Dana : SourceType("dana")
+    object Gcash : SourceType("gcash")
+    object Kakaopay : SourceType("kakaopay")
+    object TouchNGo : SourceType("touch_n_go")
+    object Fpx : SourceType("fpx") {
+        var banks : List<Bank>? = null
+    }
     data class Unknown(override val name: String?) : SourceType(name)
 
     sealed class InternetBanking(@JsonValue override val name: String?) : SourceType(name) {
@@ -35,6 +44,9 @@ sealed class SourceType(
     }
 
     sealed class MobileBanking(@JsonValue override val name: String?) : SourceType(name) {
+        object Bay : MobileBanking("mobile_banking_bay")
+        object KBank : MobileBanking("mobile_banking_kbank")
+        object OcbcPao : MobileBanking("mobile_banking_ocbc_pao")
         object Scb : MobileBanking("mobile_banking_scb")
         data class Unknown(@JsonValue override val name: String?) : MobileBanking(name)
     }
@@ -43,9 +55,13 @@ sealed class SourceType(
         object Bay : Installment("installment_bay")
         object FirstChoice : Installment("installment_first_choice")
         object Bbl : Installment("installment_bbl")
+        object Ezypay : Installment("installment_ezypay")
         object Ktc : Installment("installment_ktc")
         object KBank : Installment("installment_kbank")
         object Scb : Installment("installment_scb")
+        object Citi : Installment("installment_citi")
+        object Ttb : Installment("installment_ttb")
+        object Uob : Installment("installment_uob")
         data class Unknown(@JsonValue override val name: String?) : Installment(name)
 
         companion object {
@@ -54,9 +70,13 @@ sealed class SourceType(
                         Bay -> listOf(3, 4, 6, 9, 10)
                         FirstChoice -> listOf(3, 4, 6, 9, 10, 12, 18, 24, 36)
                         Bbl -> listOf(4, 6, 8, 9, 10)
+                        Ezypay -> listOf(6, 12, 24)
                         Ktc -> listOf(3, 4, 5, 6, 7, 8, 9, 10)
                         KBank -> listOf(3, 4, 6, 10)
                         Scb -> listOf(3, 4, 6, 9, 10)
+                        Citi -> listOf(4, 5, 6, 7, 8, 9, 10)
+                        Ttb -> listOf(3, 4, 5, 6, 7, 8, 9, 10)
+                        Uob -> listOf(4, 5, 6, 7, 8, 9, 10)
                         is Unknown -> emptyList()
                     }
         }
@@ -71,21 +91,35 @@ sealed class SourceType(
             "internet_banking_ktb" -> InternetBanking.Ktb
             "internet_banking_scb" -> InternetBanking.Scb
             "internet_banking_bbl" -> InternetBanking.Bbl
+            "mobile_banking_bay" -> MobileBanking.Bay
+            "mobile_banking_kbank" -> MobileBanking.KBank
+            "mobile_banking_ocbc_pao" -> MobileBanking.OcbcPao
             "mobile_banking_scb" -> MobileBanking.Scb
             "alipay" -> Alipay
             "bill_payment_tesco_lotus" -> BillPaymentTescoLotus
             "barcode_alipay" -> BarcodeAlipay
             "econtext" -> Econtext
+            "fpx" -> Fpx
             "truemoney" -> TrueMoney
             "installment_bay" -> Installment.Bay
             "installment_first_choice" -> Installment.FirstChoice
             "installment_bbl" -> Installment.Bbl
+            "installment_ezypay" -> Installment.Ezypay
             "installment_ktc" -> Installment.Ktc
             "installment_kbank" -> Installment.KBank
             "installment_scb" -> Installment.Scb
+            "installment_citi" -> Installment.Citi
+            "installment_ttb" -> Installment.Ttb
+            "installment_uob" -> Installment.Uob
             "points_citi" -> PointsCiti
             "paynow" -> PayNow
             "promptpay" -> PromptPay
+            "alipay_cn" -> AlipayCn
+            "alipay_hk" -> AlipayHk
+            "dana" -> Dana
+            "gcash" -> Gcash
+            "kakaopay" -> Kakaopay
+            "touch_n_go" -> TouchNGo
             else -> Unknown(name)
         }
     }
@@ -111,14 +145,19 @@ val SourceType.Companion.allElements: List<SourceType>
             SourceType.BillPaymentTescoLotus,
             SourceType.BarcodeAlipay,
             SourceType.Econtext,
+            SourceType.Fpx,
             SourceType.TrueMoney,
             SourceType.Installment.Bay,
             SourceType.Installment.FirstChoice,
             SourceType.Installment.Bbl,
+            SourceType.Installment.Ezypay,
             SourceType.Installment.Ktc,
             SourceType.Installment.KBank,
             SourceType.PointsCiti,
-            SourceType.Installment.Scb
+            SourceType.Installment.Scb,
+            SourceType.Installment.Citi,
+            SourceType.Installment.Ttb,
+            SourceType.Installment.Uob
 
     )