+++ sdk/src/main/java/co/omise/android/ui/PaymentChooserFragment.kt
@@ -43,7 +43,14 @@ internal class PaymentChooserFragment : OmiseListFragment<PaymentMethodResource>
             PaymentMethodResource.Alipay,
             PaymentMethodResource.PayNow,
             PaymentMethodResource.PromptPay,
+            PaymentMethodResource.AlipayCn,
+            PaymentMethodResource.AlipayHk,
+            PaymentMethodResource.Dana,
+            PaymentMethodResource.Gcash,
+            PaymentMethodResource.Kakaopay,
+            PaymentMethodResource.TouchNGo,
             PaymentMethodResource.PointsCiti -> item.sourceType?.let(::sendRequest)
+            PaymentMethodResource.Fpx -> navigation.navigateToFpxEmailForm()
             PaymentMethodResource.GooglePay -> navigation.navigateToGooglePayForm()// need to change
         }
     }