+++ sdk/src/sharedTest/java/co/omise/android/ui/InstallmentChooserFragmentTest.kt
@@ -2,6 +2,7 @@ package co.omise.android.ui
 
 import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.hasDescendant
@@ -25,10 +26,14 @@ class InstallmentChooserFragmentTest {
     private val paymentMethods = listOf(
             PaymentMethod(name = "installment_bay"),
             PaymentMethod(name = "installment_bbl"),
+            PaymentMethod(name = "installment_ezypay"),
             PaymentMethod(name = "installment_first_choice"),
             PaymentMethod(name = "installment_kbank"),
             PaymentMethod(name = "installment_ktc"),
-            PaymentMethod(name = "installment_scb")
+            PaymentMethod(name = "installment_scb"),
+            PaymentMethod(name = "installment_citi"),
+            PaymentMethod(name = "installment_ttb"),
+            PaymentMethod(name = "installment_uob")
     )
     private val mockNavigation: PaymentCreatorNavigation = mock()
     private val fragment = InstallmentChooserFragment.newInstance(paymentMethods).apply {
@@ -45,10 +50,18 @@ class InstallmentChooserFragmentTest {
     fun displayAllowedInstallmentBanks_showAllowedInstallmentBanksFromArgument() {
         onView(withListId(R.id.recycler_view).atPosition(0)).check(matches(hasDescendant(withText(R.string.payment_method_installment_bay_title))))
         onView(withListId(R.id.recycler_view).atPosition(1)).check(matches(hasDescendant(withText(R.string.payment_method_installment_bbl_title))))
-        onView(withListId(R.id.recycler_view).atPosition(2)).check(matches(hasDescendant(withText(R.string.payment_method_installment_first_choice_title))))
-        onView(withListId(R.id.recycler_view).atPosition(3)).check(matches(hasDescendant(withText(R.string.payment_method_installment_kasikorn_title))))
-        onView(withListId(R.id.recycler_view).atPosition(4)).check(matches(hasDescendant(withText(R.string.payment_method_installment_ktc_title))))
-        onView(withListId(R.id.recycler_view).atPosition(5)).check(matches(hasDescendant(withText(R.string.payment_method_installment_scb_title))))
+        onView(withListId(R.id.recycler_view).atPosition(2)).check(matches(hasDescendant(withText(R.string.payment_method_installment_ezypay_title))))
+        onView(withListId(R.id.recycler_view).atPosition(3)).check(matches(hasDescendant(withText(R.string.payment_method_installment_first_choice_title))))
+        onView(withListId(R.id.recycler_view).atPosition(4)).check(matches(hasDescendant(withText(R.string.payment_method_installment_kasikorn_title))))
+        onView(withListId(R.id.recycler_view).atPosition(5)).check(matches(hasDescendant(withText(R.string.payment_method_installment_ktc_title))))
+        onView(withListId(R.id.recycler_view).atPosition(6)).check(matches(hasDescendant(withText(R.string.payment_method_installment_scb_title))))
+
+        onView(withId(R.id.recycler_view)).perform(ViewActions.swipeUp());
+
+        onView(withListId(R.id.recycler_view).atPosition(7)).check(matches(hasDescendant(withText(R.string.payment_method_installment_citi_title))))
+        onView(withListId(R.id.recycler_view).atPosition(8)).check(matches(hasDescendant(withText(R.string.payment_method_installment_ttb_title))))
+        onView(withListId(R.id.recycler_view).atPosition(9)).check(matches(hasDescendant(withText(R.string.payment_method_installment_uob_title))))
+
         onView(withId(R.id.recycler_view)).check(matches(itemCount(paymentMethods.size)))
     }